new dao funcs

+++ Brief_10/src/servlet/Articles.java
@@ -22,27 +22,50 @@ import java.util.Calendar;
 public class Articles extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         final String _action = request.getParameter("action");
+        final int idArticle = Integer.parseInt(request.getParameter("article"));
+        Utilisateur user = (Utilisateur) request.getSession(false).getAttribute("__user_data");
+        //
         String ret_data = "";
         switch (_action) {
             case "vote":
-                final int idArticle = Integer.parseInt(request.getParameter("article"));
-                Utilisateur user = (Utilisateur) request.getSession(false).getAttribute("__user_data");
-                try {
-                    VoteDao voteDao = new VoteDao();
-                    Vote vote = voteDao.get(idArticle, user.getIdutilisateur());
-                    if (vote == null) {
-                        boolean insertRes = voteDao.insert(new Vote(idArticle, user.getIdutilisateur()));
-                        if (insertRes) ret_data = "{\"status\":1}";
-                        else ret_data = "{\"status\":0}";
-                    } else{
-                        boolean removeRes = voteDao.delete(new Vote(idArticle, user.getIdutilisateur()));
-                        if (removeRes) ret_data = "{\"status\":1}";
-                        else ret_data = "{\"status\":0}";
+                if(user.getRoleUtilisateur().equals("client")) {
+                    try {
+                        VoteDao voteDao = new VoteDao();
+                        Vote vote = voteDao.get(idArticle, user.getIdutilisateur());
+                        if (vote == null) {
+                            boolean insertRes = voteDao.insert(new Vote(idArticle, user.getIdutilisateur()));
+                            if (insertRes) ret_data = "{\"status\":1}";
+                            else ret_data = "{\"status\":0}";
+                        } else {
+                            boolean removeRes = voteDao.delete(new Vote(idArticle, user.getIdutilisateur()));
+                            if (removeRes) ret_data = "{\"status\":1}";
+                            else ret_data = "{\"status\":0}";
+                        }
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                        ret_data = "null";
                     }
-                } catch (Exception e) {
-                    e.printStackTrace();
-                    ret_data = "null";
-                }
+                }else ret_data = "{\"status\":-1}";
+                break;
+            case "delete":
+                if(user.getRoleUtilisateur().equals("admin")) {
+                    try {
+                        VoteDao voteDao = new VoteDao();
+                        Vote vote = voteDao.get(idArticle, user.getIdutilisateur());
+                        if (vote == null) {
+                            boolean insertRes = voteDao.insert(new Vote(idArticle, user.getIdutilisateur()));
+                            if (insertRes) ret_data = "{\"status\":1}";
+                            else ret_data = "{\"status\":0}";
+                        } else {
+                            boolean removeRes = voteDao.delete(new Vote(idArticle, user.getIdutilisateur()));
+                            if (removeRes) ret_data = "{\"status\":1}";
+                            else ret_data = "{\"status\":0}";
+                        }
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                        ret_data = "null";
+                    }
+                }else ret_data = "{\"status\":-1}";
                 break;
         }
         //
